work-in-progress to support deletion using an idLookup field

+++ src/main/java/com/salesforce/dataloader/action/DeleteAction.java
@@ -30,6 +30,8 @@ package com.salesforce.dataloader.action;
 import com.salesforce.dataloader.action.progress.ILoaderProgress;
 import com.salesforce.dataloader.action.visitor.DAOLoadVisitor;
 import com.salesforce.dataloader.action.visitor.DeleteVisitor;
+import com.salesforce.dataloader.action.visitor.RESTDeleteVisitor;
+import com.salesforce.dataloader.config.Config;
 import com.salesforce.dataloader.controller.Controller;
 import com.salesforce.dataloader.exception.DataAccessObjectException;
 
@@ -44,7 +46,12 @@ class DeleteAction extends AbstractLoadAction {
 
     @Override
     protected DAOLoadVisitor createVisitor() {
-        return new DeleteVisitor(getController(), getMonitor(), getSuccessWriter(), getErrorWriter());
+        if (getController().getConfig().isRESTAPIEnabled()
+                && getController().getConfig().getBoolean(Config.DELETE_WITH_EXTERNALID)) {
+            return new RESTDeleteVisitor(getController(), getMonitor(), getSuccessWriter(), getErrorWriter());
+        } else {
+            return new DeleteVisitor(getController(), getMonitor(), getSuccessWriter(), getErrorWriter());
+        }
     }
 
 }
